### 2016-10-27 #98 @whedon
[PRE REVIEW]: acronym: An Automatic Reduction Pipeline for Astronomical Images
**Submitting author:** @kweis (<a href="http://orcid.org/0000-0002-4639-1436">Kolby Weisenburger</a>)
**Repository:** <a href="https://github.com/kweis/acronym" target ="_blank">https://github.com/kweis/acronym</a>
**Version:** v1.0.0
