### 2016-10-27 #102 @arfon
@xuanxu :wave:. Many thanks for agreeing to review this submission. Please follow the reviewer guidelines (http://joss.theoj.org/about#reviewer_guidelines) and update the issue checklist at the top of this issue as you go.
