### 2016-10-30 kweis/acronym#5 @migueldvb
import either astropy.io.fits or pyfits 
`pyfits` is currently imported from two different places in acronym.py. Since the pyfits standalone version is deprecated in favor of astropy and have the same functionality it should be tried first to import the `astropy.io.fits` package and otherwise fall to standalone pyfits.

```
try:
    import astropy.io.fits as pyfits
except ImportError:
    import pyfits
```
