### 2016-10-31 kweis/acronym#6 @migueldvb
Use os.path.join to merge path components
For example.

```
    os.makedirs(os.path.join(direc, 'reduced', 'cals'))
```

instead of

```
    os.makedirs(direc+'/reduced/cals')
```

This gets rid of the extra slashes and is portable.
