### 2016-10-31 kweis/acronym#7 @migueldvb
python 3 compatibility
It is not clear from the documentation if the code is python 3 compatible. It looks that the only change needed is to convert the iterators generated by the filter function in lines 131 and 153 to lists.

```
times = list(filter(None,pd.unique(df.exp.ravel())))
```

```
filters = list(filter(None,pd.unique(df.filt.ravel())))
```
