### 2016-10-31 kweis/acronym#8 @migueldvb
FileNotFoundError
When the input data directory is given with a trailing slash (as it happens with bash and zsh tab completion) the script fails with the following error:

```
python acronym.py example/rawdata/
```

```
Traceback (most recent call last):
  File "acronym.py", line 210, in <module>
    pyfits.writeto(name,dat,clobber=True,header=dat_head)
  File "/usr/lib/python3.5/site-packages/astropy/io/fits/convenience.py", line 424, in writeto
    checksum=checksum)
  File "/usr/lib/python3.5/site-packages/astropy/io/fits/hdu/base.py", line 462, in writeto
    checksum=checksum)
  File "/usr/lib/python3.5/site-packages/astropy/io/fits/hdu/hdulist.py", line 686, in writeto
    fileobj = _File(fileobj, mode='ostream', clobber=clobber)
  File "/usr/lib/python3.5/site-packages/astropy/io/fits/file.py", line 150, in __init__
    self._open_filename(fileobj, mode, clobber)
  File "/usr/lib/python3.5/site-packages/astropy/io/fits/file.py", line 502, in _open_filename
    self._file = fileobj_open(self.name, PYFITS_MODES[mode])
  File "/usr/lib/python3.5/site-packages/astropy/io/fits/util.py", line 389, in fileobj_open
    return open(filename, mode, buffering=0)
FileNotFoundError: [Errno 2] No such file or directory: 'example/rawdata//reduced/data/red_example/rawdata/m33_V_120.0051.fits'
```

This can be fixed by defining the output FITS file in line 210 using the os.path.basename function.

```
    name = direc+'/reduced/data/red_'+(df['fname'][n]).replace(direc+"/","")
```

replaced by:

```
    name = os.path.join(direc, 'reduced/data', 'red_'+os.path.basename(df['fname'][n]))
```
