### 2016-10-31 kweis/acronym#9 @migueldvb
Remove matplotlib dependency
Since matplotlib is not used in `acronym.py` the two imports of `matplotlib.pyplot` and `pylab` near the top of the file should be removed.
