### 2016-10-31 kweis/acronym#10 @migueldvb
Large binary files in repository
There are about 140 MB of data files and some deleted ones that are in the commit history that are used to test the code. These files could be downloaded from a test function (if needed) to speed up the process of cloning the repository for users that do not need the data files.
