### 2016-10-31 kweis/acronym#11 @migueldvb
version number not available
The JOSS submission https://github.com/openjournals/joss-reviews/issues/102 lists version 1.0.0 of the code. However, the github repository does not have any release tags.
