### 2016-11-01 #102 @migueldvb
As a general comment it seems that this CCD reduction could be done with
existing open source software such as astropy's ccdproc package (implemented in
python) or FITSH using some helper functions from acronymn (as ARCTIC data have
some FITS keywords READAMPS, DSEC11 that are not part of the FITS standard
such as TRIMSEC and BIASSEC).  However I may have missed something.  It would be
useful to say in the documentation what is different in this code from other
reduction software and give a reference to other packages that have
the same functionality.

Another issue that should be fixed is that currently there is no automated test
functions that will verify that the reduction of the test data in he repository
provides the expected result.

There are a few additional minor issues in the acronym github repository (some of them may have already been fixed).

- https://github.com/kweis/acronym/issues/5
- https://github.com/kweis/acronym/issues/6
- https://github.com/kweis/acronym/issues/6
- https://github.com/kweis/acronym/issues/7
- https://github.com/kweis/acronym/issues/8
- https://github.com/kweis/acronym/issues/9
- https://github.com/kweis/acronym/issues/10
- https://github.com/kweis/acronym/issues/11

So far I haven't got any other comments. If I find any other issues I will open a new ticket.
