### 2016-11-01 9211e00
addressing issues opened by @migueldvb
