### 2016-11-05 3619dd8
moved example dir
