### 2016-12-12 #102 @arfon
> @kweis - the paper.md won't compile as there aren't any references? Given you mention Astropy in your readme perhaps you should consider (at least) citing this package?

Friendly bump @kweis.
