### 2017-01-20 #102 @sagetrey
unsubscribe

On Fri, Jan 20, 2017 at 9:35 AM, Arfon Smith <notifications@github.com>
wrote:

> 👋 @kweis <https://github.com/kweis>
>
> Hi @kweis <https://github.com/kweis> - taking a look at the list of
> issues that @migueldvb <https://github.com/migueldvb> opened on your
> project it appears that you closed these with no comment:
>
>    - kweis/acronym#5 <https://github.com/kweis/acronym/issues/5>
>    - kweis/acronym#6 <https://github.com/kweis/acronym/issues/6>
>    - kweis/acronym#7 <https://github.com/kweis/acronym/issues/7>
>    - kweis/acronym#8 <https://github.com/kweis/acronym/issues/8>
>    - kweis/acronym#9 <https://github.com/kweis/acronym/issues/9>
>    - kweis/acronym#10 <https://github.com/kweis/acronym/issues/10>
>    - kweis/acronym#11 <https://github.com/kweis/acronym/issues/11>
>
> It's not clear if you have addressed @migueldvb
> <https://github.com/migueldvb>'s feedback on these issues or aren't
> planning to. If it's the latter then I'm going to assume that you aren't
> interested in having your paper published in JOSS any longer.
>
> Please respond here to let us know what you'd like to do at this stage but
> if there's no response within the next couple of weeks I'll reject this
> submission.
>
