### 2017-01-20 #102 @arfon
@sagetrey - please take a look at the instructions here to unsubscribe from JOSS reviews: https://github.com/openjournals/joss-reviews/issues/166#issuecomment-273498829
