### 2017-01-23 #102 @kweis
I've resolved all opened issues and moved the large binary files to a zipped file in the release docs. I released an updated [package](https://github.com/kweis/acronym/releases/latest) that highlights how I've addressed @migueldvb's feedback.

With respect to the general comment regarding the CCD reduction -- the reason for writing in-house procedures (rather than astropy's ccdproc, for instance) is for clarity and transparency when handling ARCTIC's peculiarities. ARCTIC has specialized readout modes that result in variable image sizes. How we handle this is most obvious in the trim_image function and while this could be enveloped into a helper function in tandem with astropy's ccdproc, we wished to maintain a package that was curated specifically for the ARCTIC instrument and that did not rely on preexisting reduction packages. We are currently comparing our results to results from using ccdproc and will update the repository with a test function this week.
