### 2017-02-05 #102 @arfon
> We are currently comparing our results to results from using ccdproc and will update the repository with a test function this week.

👋 @kweis - how are you getting on with this update?
