### 2017-03-06 kweis/acronym#12 @migueldvb
replace deprecated "clobber=True" argument by "overwrite=True"
Since astropy 1.3 the clobber argument in `astropy.io.fits.writeto` has been replaced by overwrite.

````
WARNING: AstropyDeprecationWarning: "clobber" was deprecated in version 1.3 and will be removed in a future version. Use argument "overwrite" instead. [astropy.utils.decorators]
 ````

This issue is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/102
