### 2017-03-06 #102 @migueldvb
Thanks @kweis I have updated the issue checklist at the top of the review with the current status of the package. I will be glad to review the updated version when it is ready.
