### 2017-03-31 kweis/acronym#12 @kweis
Replaced clobber with overwrite. Thanks
