### 2017-03-31 ac53043
adding references, updating deprecated functions
