+++ requirements.txt
@@ -1,0 +1,3 @@
+numpy>=1.12.1
+pandas>=0.15.2
+astropy>=1.3.2
