### 2017-04-14 #102 @kweis
Hi there,
I've addressed the outstanding checklist items as follows:
1. Installation instructions are in the README and I've added a requirements.txt file to install/update the required dependencies. Since there are only three nonstandard libraries, I believe the requirements.txt file handles the dependencies most succinctly.
2. Documentation was added to the major functions and clarifying comments were added throughout the code.
3. I've included appropriate references in the README and in the paper.md file.
