### 2017-04-20 #102 @migueldvb
Thank you for the revised version. I will send you any further comments by the end of the week, if I have any.
