### 2017-04-24 #102 @migueldvb
I have a couple of comments on the latest version.  The README file refers to  "/example/rawdata zipfile (example.zip in release docs)". However there is no such file in the `docs` directory and I could not figure out how to download the test data from the description in the README. Since the package depends on astropy already I'd suggest using `astropy.utils.data` tor download the test data but this could be done in a number of other ways.  There is another reference in the README to the "use the test.py script (see below)" but this script cannot be found in the repo. It may have been just an oversight in not adding this file. Finally the suggested use of os.path for path manipulation is included in https://github.com/kweis/acronym/pull/13 that is currently open.

Apart from those issues the requirements for acceptance are met now and I think it is almost ready.
