### 2017-05-02 #102 @kweis
I cleaned up the README to address the most recent comments from @migueldvb -- the raw data is still in the release notes to keep it out of the main repository, but I linked to it so it's much easier to find.

I think I'd already addressed the os.path.join issue, but forgot to close it. closed now!
