### 2017-05-02 93eeeb6
cleaned up README
