### 2017-05-02 2004a1f
path manipulation
