+++ acronym.py
@@ -38,16 +38,18 @@ if len(sys.argv) > 1:
 else:
     direc = '.'
 
-    
+cals_direc = os.path.join(direc, 'reduced', 'cals')
+reduced_direc = os.path.join(direc, 'reduced', 'data')
+
 # directories for reduced images
-if not os.path.exists(direc+'/reduced/cals'):
-    os.makedirs(os.path.join(direc, 'reduced/cals'))
-if not os.path.exists(direc+'/reduced/data'):
-    os.makedirs(os.path.join(direc, 'reduced/data'))
+if not os.path.exists(cals_direc):
+    os.makedirs(cals_direc)
+if not os.path.exists(reduced_direc):
+    os.makedirs(reduced_direc)
 
     
 # grab all files from the directory; organize dataframe
-files = glob.glob(direc+"/*.fits")
+files = glob.glob(os.path.join(direc, "*.fits"))
 
 df = pd.DataFrame(files,columns=['fname'])
 df['objtype'] = pd.Series("", index=df.index)
@@ -123,12 +125,11 @@ def getdark(expt):
     """
     
     try:
-        dark = pyfits.getdata(direc+'/reduced/cals/master_dark_'+str(expt)+'.fits')
+        dark = pyfits.getdata(os.path.join(cals_direc,'master_dark_{0}.fits'.format(expt)))
     except IOError:
         scaleto = np.max(df['exp'][df['exp'] != ''])
-        dark = pyfits.getdata(direc+'/reduced/cals/master_dark_'+str(scaleto)+'.fits')
+        dark = pyfits.getdata(os.path.join(cals_direc,'master_dark_{0}.fits'.format(scaleto)))
         dark *= (expt/scaleto)
-
     return dark
 
 
@@ -144,7 +145,7 @@ if len(bias_idx) == 0:
 else:
     biases = np.array([trim_image(df['fname'][n])[0] for n in bias_idx])
     bias = np.median(biases,axis=0)
-    pyfits.writeto(direc+'/reduced/cals/master_bias.fits',bias,overwrite=True)
+    pyfits.writeto(os.path.join(cals_direc, 'master_bias.fits'),bias,overwrite=True)
     print('   > Created master bias')
 
 
@@ -164,8 +165,8 @@ for ii in range(0,len(times)):
     else:
         darks = np.array([trim_image(df['fname'][n])[0] for n in dark_idx]) - bias
         dark_final = np.median(darks,axis=0)
-    
-        name = direc+'/reduced/cals/master_dark_'+str(times[ii])+'.fits'
+
+        name = os.path.join(cals_direc,'master_dark_{0}.fits'.format(times[ii]))
         pyfits.writeto(name,dark_final,overwrite=True)
         print('   > Created master '+ str(times[ii])+' second dark')
 
@@ -198,7 +199,7 @@ for ii in range(0,len(filters)):
         flat_final /= np.max(flat_final)
 
         filts = filters[ii][-1]
-        name = direc+'/reduced/cals/master_flat_'+filts+'.fits'
+        name = os.path.join(cals_direc, 'master_flat_{0}.fits'.format(filts))
         pyfits.writeto(name,flat_final,overwrite=True)
         print('   > Created master '+ str(filters[ii])+' flat')
 
@@ -223,13 +224,13 @@ for n in dat_idx:
     
     filt = (df['filt'][n])[-1]
     try:
-        flat = pyfits.getdata(direc+'/reduced/cals/master_flat_'+str(filt)+'.fits')
+        flat = flat = pyfits.getdata(os.path.join(cals_direc,'master_flat_{0}.fits'.format(filt)))
     except IOError:
[…]
