### 2017-05-02 #102 @migueldvb
That sounds great. I recommend this paper for acceptance in JOSS.
