### 2017-05-02 #102 @arfon
> That sounds great. I recommend this paper for acceptance in JOSS.

Great!
