+++ docs/paper.bib
@@ -1,0 +1,34 @@
+@misc{ccdproc,
+  author       = {Matt Craig and
+                  Steve Crawford and
+                  Thomas Robitaille and
+                  Brigitta Sipocz and
+                  Joe Philip Ninan and
+                  Michael Droettboom and
+                  Erik Tollerud and
+                  Erik Bray and
+                  Michael Seifert and
+                  walkerna22 and
+                  Hans Moritz Günther and
+                  stottsco and
+                  Evert Rol and
+                  Larry Bradley and
+                  Christoph Deil and
+                  Anthony Horton and
+                  Kyle Barbary and
+                  Nathan and
+                  Forrest Gasdia and
+                  Sourav Singh and
+                  Stefan Nelson and
+                  Wolfgang Kerzendorf and
+                  William Schoenell and
+                  Punyaslok Pattnaik and
+                  Ole Streicher and
+                  Carlos Gomez and
+                  Benjamin Alan Weaver},
+  title        = {ccdproc: v1.0},
+  month        = mar,
+  year         = 2016,
+  doi          = {10.5281/zenodo.47652},
+  url          = {https://doi.org/10.5281/zenodo.47652}
+}
