### 2017-05-02 #102 @kweis
Okay, moved it to the paper.bib file and cited using @ in the paper.md. Is that the citation you were referring to?
