### 2017-05-02 #102 @arfon
> Okay, moved it to the paper.bib file and cited using @ in the paper.md. Is that the citation you were referring to?

Yep. That looks good.

@migueldvb - many thanks for your review here ✨

@kweis - your submission is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00102 ⚡️ 🚀 💥
